pi-client/requirements.txt
numpy>=1.26,<3
opencv-python-headless>=4.10,<5
python-dotenv>=1.0,<2
requests>=2.32,<3
sounddevice>=0.4,<1
soundfile>=0.12,<1

